move fetchGroups to app layout so state is always loaded
