+++ expo/components/ChatSelect/ChatSelect.tsx
@@ -9,45 +9,11 @@ import { CanceledError } from "axios";
 export const ChatSelect = () => {
   const { getGroups } = useWebSocket();
   const { store, user, groups, setGroups } = useGlobalStore();
-
-  const isFetching = useRef(false);
-
-  const fetchGroups = async () => {
-    if (isFetching.current) return;
-    isFetching.current = true;
-
-    try {
-      const data = await getGroups();
-      setGroups(data);
-      store.saveGroups(data);
-    } catch (error) {
-      if (!(error instanceof CanceledError)) {
-        try {
-          const storedGroups = await store.loadGroups();
-          setGroups(storedGroups);
-        } catch (storeError) {
-          console.error("Failed to load groups:", storeError);
-        }
-      }
-    } finally {
-      isFetching.current = false;
-    }
-  };
-
-  useEffect(() => {
-    fetchGroups();
-
-    const intervalId = setInterval(fetchGroups, 20000);
-
-    return () => clearInterval(intervalId);
-  }, []);
-
   return (
     <View style={styles.container}>
       {user ? (
         <View>
           <ChatCreate user={user} />
-          <Button onPress={fetchGroups} title={"Refresh"} />
           {groups.map((group, index) => {
             return (
               <ChatSelectBox
